Simplified filtering and naming
I changed the function call from lapply to purrr::map since I think that the map functions in that package are a bit more user friendly, and their arguments are more consistent across functions. Also, both purrr::map and lapply will use the names of the input on the output, so that's why I use setNames in the call to purrr::map in order to not have to set the names later.