Use purrr::map instead of lapply and bind_rows
In addition, the original way of formulating things can be dangerous. When collecting the results from the Shapiro-Wilks test, the column `Shapiro_OK` is set to either `"TRUE"` or `"FALSE"`, i.e. character values, not booleans. In the comparison that then is made in order to fish out things that needs to be normalised, this character column is strictly compared to `FALSE`. I didn't expect this to work at all, since I wouldn't consider `FALSE == "FALSE"` to be true, but apparently it is.

Anyway, I would recommend to just use booleans in order to minimise confusion and possible hard-to-track errors.