Catch warnings on first iteration
The old code did the normalisation twice in order to catch any warnings that were generated. This version only normalises the columns once and catches warnings at the same time. The printing might not be as pretty as before, but I will argue that it is more semantically correct to use `warning` instead of `cat`.