Refactor BLUP calculation
The calculation itself is untouched. What I did was to pull out the NaN/Inf checking and doing that on all trait columns before building models for the individual traits. This should be more efficient in the long run.

I also merged the BLUP calculation with the generation of the final BLUP table, so this should save some extra iterations and memory.