Add back outlier_dat variable
In my opinion, this data frame is not really needed. This information is really contained in the outliers_residuals object which contains the rows that are considered outliers for each variable. If you want the values explicitly, then I would say use that variable downstream instead of saving yet another data frame in the results.