Make pipe function readily available
Much of the function depends on dplyr, and in most cases the functions from dplyr are prefixed with `dplyr::`, but this is not the case for the pipe (`%>%`). By adding this, we still don't have to call `library(dplyr)`, and we don't have to prefix it with `dplyr::` all the time.